Update patch set 13
Patch Set 13:

I have dig into it and it is not so trivial to do it cleanly. Since MediaWiki install.php detects skins and add a wfLoadSkin() automatically, I will add support to inject additional repositories to clone in the prepare-mediawiki-zuul-project macro, this way it will not inject a skin in deps.txt / extensions_load.txt which would be easier to handle.

Can look at coming up with a better system later on.

Patch-set: 13
Label: Verified=0