nodepool: delete DHCP leases from snapshot
The snapshot image has a DHCP lease associated. When we later boot
instances out of it, the DHCP client first tries to retrieve the same
lease by issuing a few DHCPREQUEST before falling back to DHCPDISCOVER.

The DHCPREQUEST delay the boot by roughly 30 seconds and are completely
unnecessary.

Delete the DHCP leases from the instance to be snapshotted. Tested the
sequence manually and confirmed it work.

Bug: T113342
Change-Id: Iecb880d946cba00b8d9a8b5fa4e4313c0a5ac83b