Update patch set 18
Patch Set 18:

So that change is rather terrible but should work, a demo is the child change https://gerrit.wikimedia.org/r/#/c/326469/

I am injecting the skins in their own parameter variable. The reason is we want to clone them just like skins but skins should not end in mediawiki_loads.txt, that is the MediaWiki installer that takes care of adding them in LocalSettings.php

I think it is going to work fine.

Slightly related:
* The selenium tests jobs will require Vector to be injected (will be for another change though
* Wikibase jobs use a different code path based on JJB variables. Ditto, for another change.

Patch-set: 18