Reuse macro in mediawiki-core-selenium-jessie
The job copy pasted logic from prepare-mediawiki-zuul-project, that has
been done originally because we did not support injecting skins which is
now possible.

Replace the duplicate code with the macro prepare-mediawiki-zuul-project

Set Vector as a skin dependency by passing a parameter directly in the
job, the reason is the Zuul parameter_functions does not trigger on
mediawiki/core.

Add the job on mediawiki/skins/Vector as an experimental one. Later on
we will want the job to be shared on both core and Vector to prevent
Vector from causing core to fail.

Update: mediawiki-core-selenium-jessie

Bug: T139740
Change-Id: I24cf2aaa51d9ed7634fcdc046a637792e95f10a8