Properly inject SKIN_DEPENDENCIES
For mediawiki-core-selenium-jessie, we need to inject the Vector skin as
a dependency via SKIN_DEPENDENCIES.  I thought about adding it directly
in the Jenkins job definition, however the Gearman plugin does not merge
Zuul parameters with the one from the job.

Remove the paramter from the job level (from 3a5661d0a) and move the
logic to the zuul parameter function.

Add tests to assert SKIN_DEPENDENCIES is injected and kept when Vector
is appended.
Since the test alter the 'dependencies' global, add some setUp/tearDown
to restore it in its original state.

Bug: T139740
Change-Id: I82db52a5c0205cf846f3945a696004fa5f19d4d0