mwext-mw-selenium now get Vector dep from Zuul
6c4d8f250a originaly implemented the mwext-mw-selenium by cloning the
Vector skin but since that is done after prepare-mediawiki-zuul-project,
the skin was not injected in LocalSettings.php on the first run. On
permanent slaves since the workspace is kept between build, the next
build will pass but on Nodepool that would fail (T137112).

Reuse the logic from previous changes to have Zuul inject Vector has a
dependency. Thus prepare-mediawiki-zuul-project clones it and the
MediaWiki installer then inject the skin.

Update
    mwext-mw-selenium
    mwext-mw-selenium-composer
    mwext-mw-selenium-composer-jessie
    mwext-mw-selenium-jessie

Bug: T137112
Change-Id: Ic7c83638b7cee64914c4a158bcde3e0e4c2bc88c