Update patch set 4
Patch Set 4: Code-Review-1

Thank you a lot for your effort, but I fear unfortunately this is not exactly what we want. This will send emails on each failure for the Wikibase repository (which is the active development repository). This would cause a lot email every day and would simply be too much.
We only want to be notified on failures for the Wikidata-Build repository, where a new build is pushed automatically once a day.
I've also updated the related task in T152495 to be more precise.

Patch-set: 4
Label: Code-Review=-1