Ensure postmerge reports to Gerrit with no votes
We had the postmerge pipeline set with verified=0 to force Zuul to
report (lack of parameters used to make it to not report).  Since Gerrit
2.13, we can no more vote --verified=0 on a closed change (T153737).

Pass an empty dictionary of parameters on postmerge success/failure just
like we did for the experimental pipeline.

Add test to ensure all pipelines (except post/publish) are reporting to
Gerrit on both success and failure.

Test the 'gerrit review' command crafted or the postmerge pipeline. If
that breaks again, that let us define what is expected by Gerrit and
easily adjust the parameters.

Bug: T153737
Change-Id: I25d7f7bc458fad2db46354d1466955748d6877fe