Update patch set 2
Patch Set 2:

Made the gerrit success/failure to pass an empty dictionary. That is what we have done on experimental pipeline.

Polished up the test that ensure every pipeline are reporting to Gerrit on success/failure.

Add a lame test that verify the 'gerrit review' command forged for the postmerge pipeline.  Will have to test it out for real and if it reports properly on a closed change we can land this.

Patch-set: 2
Label: Verified=0