Update patch set 1
Patch Set 1:

having a separate javadoc config for the highlighter sounds good to me. Moreover because it requires javadoc:aggregate instead of javadoc:javadoc. Note that javadoc:aggregate seems to work on mono-module projects...

having a single job sounds go to me as well. This leaves the problem of the post merge build where we need to redo everything.

Patch-set: 1
Label: Verified=0