Update patch set 10
Abandoned

That followed a discussion Paladox and I had, merely let people comment in Gerrit:

 compile foobar.eqiad.wmnet

And that would pass the parameters to the jenkins job.

Turns out Zuul does not pass the event which is only used to enqueue a change in one of the pipeline/queue.  When the jobs are launched, the context of the event is NOT attached so we can really use that.

Patch-set: 10
Status: abandoned