Fix chromium browser detection on jessie
Keeps backwords support for trusty but this fixes support for jessie.

Chromium's package name is different on jessie then it is on trusty, so we need to export the different name otherwise the which command will fail to find the path.

Follows-up to Ie9f906f896e6a316c647553af29b61161bbd8577

Bug: T153756
Change-Id: I71a3e0461c4d12db87bceffc53f7aeb9b5815107