Update patch set 1
Patch Set 1:

HierarchyBuilder Html2Wiki used the 'mw-checks' template which has job in the check-voter pipeline.  When adding 'npm' that causes jobs to be in two check pipelines:

* check
* check-voter

I have changed the template to 'mw-checks-test' which are the same set of jobs as 'mw-checks' but in the 'check' pipeline.

Patch-set: 1
Label: Verified=0