Update patch set 1
Patch Set 1:

I hope there would not be duplicates in the final dependency list (have not test that).

To see the problem, have a look at the experimental build for PollNY extension on https://gerrit.wikimedia.org/r/#/c/331501/. The test mwext-testextension-hhvm-composer is passing, but the similar test mediawiki-extensions-hhvm-jessie fails, looking at the output the first test sets the dependency to SocialeProfile, the later just set the list of gated extensions and SocialeProfile is not in that list, giving a failure. Hoping with this change both tests are passing.

Patch-set: 1
Reviewer: Umherirrender <161@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0