Update patch set 1
Patch Set 1:

Sorry I replied in a hurry earlier.  The mediawiki-extensions-* jobs are intended for Wikimedia deployed extensions.   The job is in the experimental pipeline to verify whether an extension pass and can be included in those jobs.   For non wikimedia extensions, it does not make much sense (SocialProfile is not used there).

Patch-set: 1