prepare-mediawiki-zuul-project-no-sql: Use composer to install deps
prepare-mediawiki-zuul-project-no-sql should be also installing
dependencies of extensions.

This builder is used in the job running phan checks for extensions,
and extension dependencies must be also cloned, otherwise
the analysis fails if the extension has some dependencies.

Insert 'no-vendor' in the macro name.

It is only used by the phan job: mwext-php70-phan-jessie

Change-Id: Ie080af4a6717c5df6c284118dc85408ac6f42af1