Update patch set 2
Patch Set 2:

If I remember properly that was to test Wikibase with mediawiki/vendor.  But since that macro is only used for the phan job that is all fine I guess.

Might want to rename the macro or have it inlined.  Will dig in that later today or tomorrow.

Patch-set: 2
Label: Code-Review=0