Update patch set 1
Patch Set 1:

> I am not sure how it would work. Specially the entries will be added to a more specific autoloader which is unlikely to actually be loaded.

It will work fine as soon as multiple autoloaders are allowed and just kick in aone after another (it is documented), and code in Wikibase takes care of including this autoloader. see: `lib/WikibaseLib.php:51`

> You would also end up with potential conflicts between mediawiki/core require-dev versions and the Wikibase ones

Yes we might, but it won't break production, it will be spotted even before patch is merged. I see that it is nasty, but it's not that bad.
Anyway, if you declare dependency in composer that doesn't match some other repository dependency version you will notice it only on `composer update` stage.

> I think it is easier to just add those test dependencies to mediawiki/core require-dev

I agree that it is easier, but:
* I am still not sure about this library/approach so I'd rather gave it a 2-3 month trial (to be able change interfaces easily or remove it completely, for example)
* It doesn't solve the problem, which is that we can not use dev tools core doesn't use

> That also makes Wikibase slightly different compared to others extensions and I would like to eventually use the exact same process for any extension.

I can understand that it would be great, but then the process should be very flexible or, may be,
configured by the code in the extension. Otherwise you will limit developers in their freedom.
Speaking of difference, it is already differs fom others as I see it ass soon as there is a job template `mwext-Wikibase-{kind}-tests-{dbflavor}-{phpflavor}`.

The biggest problem I see in this case is that Core dictates what in a very strict way.
If you look at composer.json dev section, you will see that almost all libraries have their version fixed.
I can imagine that makes more or less sense for production, but not for dev environment.
This way developers of each extension that depends on Core have to use only those libraries that core uses,
and it seems like unnecessary strictnes to me.

What I really think is good solution for this problem is to relax core's dependencies definition as much as possible
and merge dev dependencies. I would agree to do it, but saddly I have no idea what can break.
If someone is ready to help me and take part of responsibility, then we can work on that.

Summarizing, I see local `composer update` as a compromise.

Patch-set: 1