zuul: fix window-floor which is in changes not jobs
I originally thought the dependent pipeline window was in number of jobs
and set it to 12 to accomodate for MediaWiki.  Turns out it is in number
of changes.  So we always allowed a minimum of 12 changes to trigger
their jobs in gate-and-submit, on a failure that meant cancelling jobs
for up to 11 changes.

Make the initial window to be 5 jobs instead of the default 20.
Change window-floor from 12 to 2.  So at a minimum we will at triggers
jobs for at least two changes.

Reference:
Window parameters for pipelines on:
https://docs.openstack.org/infra/zuul/zuul.html?#pipelines

Change-Id: I4eb75af690b5f30c9f7e8ccfc6b00c5bdfcf27b8