debian-glue: only lintian against *.changes
We had the lintian utility running on the binary package via *.dsc as
well as against the source package via *.changes.  That is redundant,
the source one being a superset.

The reason is that upstream jenkins-debian-glue examples decouple the
source and binary packages builds in two jobs when our setup has
everything in the same job.

Run lintian solely againt *.changes
Change file names to lintian.{txt,xml}
Adjust Jenkins artifacts archiving and junit report to match the new
filenames.

Also dump the content of lintian.txt since we default to showing the
console.

Update:
    debian-glue
    debian-glue-non-voting

Bug: T158553
Change-Id: Ic672ac72eb0fe0aa9f2efe95f6898efa6c48922f