Delete php53lint jobs
Replace it with php55lint where relevant.
Do not replace it in check pipelines which will be phased out
eventually.

Remove the skip if from *php55* jobs. We now want to trigger Zend 5.5
jobs on REL1_23 - REL1_26 branches.

Add test suite to make sure the testextension for 5.3/5.5 are still
properly triggered. To do so add to Zuul meta job filters for php55lint
and mediawiki-core-php55lint so they get triggered for all branches.

Delete:
    mediawiki-core-php53lint
    php53lint

Bug: T158652
Change-Id: Ie6cf3de265c61cf7b5aba4c11045286fd05286f5