Add install-android-sdk step to apps-android-wikipedia-lint job
Like the other Android automated testing jobs, the Android linting job
needs a copy of the Android SDK.  It appears that the SDK for this job
(unlike the others) is currently being provided by the deprecated
application-level Android SDK Manager plugin, which we'd like to remove.

This adds the Jenkins Android Emulator Plugin install-android-sdk step
to the lint job so that we can remove the app-level SDK Manager plugin.

Bug: T147099
Change-Id: Ie71785228939e9ba514212a505e2a722966b1982