Ensure *-npm-node-6-jessie jobs get global-set-env for CHROME_BIN
This was lost in the migration from old persisent slaves (which used the
'npm' builder macro) to Nodepool.

It was only preserved by accident for mediawiki-core-qunit which
re-used the existing qunit-karma macro that runs npm-setup.sh.

This was causing I7ccaca427f1a6b in oojs/core to break when
running QUnit tests with newer karma-launcher-chrome.

Change-Id: I82fe33717dd47a3ac069ef4c6518d378364fdffe