Hygiene: Remove Jenkins plugin-driven install-android-sdk builder
The Android Gradle plugin now automatically handles SDK installation[1], so
there is no longer any need to rely on a build step powered by the Jenkins
Android emulator plugin for jobs not using the emulator.

The periodic job is using the full Gradle Android Emulator plugin and will
need to be handled separately.

Change-Id: I84ed209ef08277fdaea8cf2886e40eed91f4294f