(WIP) High priority test pipeline
Introduce a new pipeline 'test-prio' that is identical to the 'test'
pipeline but with an higher precedence.

That is done using YAML aliasing and extending test with more specific
values.

The reject-approvals.username is considered a regex and it is
precompiled.  However later Zuul does a copy.deepcopy() of the approvals
which fails because compiled regex are immutable :-/  Work around that
by dropping the jenkins-bot username and flipping the condition to cover
the cases of other users having a Verified=0 approval.

Add a quick test to ensure a CR+1 by trusted user does not trigger tests
when patch has already been tested (V -1 or V +2).

Switch operations/puppet.git from test to test-prio.

^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
WARNING
=======
A huge gotcha is whether other areas of code attempt to copy.deepcopy()
some precompiled regex.  Our test suite pass but there is no guarantee
it is not going to explode on deployment.
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

Bug: T160667
Change-Id: Iedd1b224e29ab217ca0169dc345489c763f18069