Don't ignore failures in mediawiki-core-code-coverage
We haven't these segfaults in at least years, and it's getting
rather annoying that whenever a @cover tag is misspelled, the next
Jenkins build for mediawiki-core-code-coverage ignores the failure
and publishes an empty directory.

It is a separate problem that PHPUnit considers invalid @cover
values a fatal error, but that's upstream. Right now it's a fact
that when it sees that it aborts the coverage report before
the HTML is generated (which happens after). There are no error
scenarios in which an HTML report is generated but the process
exited with an error. The segfault was an exception to this
but that is no longer the case, either.

Change-Id: Iafa3c2c0e3be3347d4c51f49bc3f1f313485e232