Update patch set 1
Patch Set 1:

Same # of instances consume.  That merely get jobs with specific name which would move the repository out of the queue of changes that has mediawiki/Wikibases etc.

Zuul merges repositories in the same queue if they are sharing a job, eg 'tox-jessie'.  by using a different job name, the repo is no more merged in the mediawiki queue.

That all got broken with the huge refactoring we did early last year :(

Patch-set: 1