Update patch set 1
Patch Set 1:

I would not disable the QUnit tests, but fix the bug instead. I can track it down to the file tests/qunit/data/mediawiki.jqueryMsg.data.js in core. It looks like it is comparing hard coded strings with translated versions, which is guaranteed to fail every time one of these translations is updated.

Since when do the failures happen? What changed? Is there a ticket for this? I could not find it but remember Jonas talking about it.

Patch-set: 1