Update patch set 1
Patch Set 1: Code-Review+1

Talked to Aleksey. We know we must re-enable the QUnit tests the moment we are adding our first relevant QUnit tests to the extension. This will take some time. currently there are none. In the meantime it is much easier to temporarily disable them, which gives us both time to continue to work on what we actually want to do (which is not JavaScript), and time to investigate the error.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1