mediawiki-core-code-coverage: Commit live hack
Follows-up 4c65c6e4 which made it so that the report is never published
because we actually have a few PHPUnit tests that normally pass but
fail when --coverage-html is enabled because it is so slow that
certain assumptions about time and clock are wrong.
(T161041, T160956, ..). These have been around for a while and
are harmless for now. Shouldn't stop us from publishing the report.

Ignore exit code, but still guard against publishing an empty directory
in case of segfaults and other fatal errors, so simply check if
index.html was created (which is the last thing before exit).

Change-Id: I2df4cbc5eb459852a83ee048f5e7d90a3aa1c789