Android: remove lint job
The -lint and -test jobs share most of their overhead and thus consume
an extra instance for little benefit.

https://gerrit.wikimedia.org/r/#/c/345121/ makes the script invoked by
the -test job to invoke lintAlphaDebug.  It already capture artifacts
matching '**/build/reports/**'.

The -lint job URL pattern points directly to the HTML build report.
However it is rather broken since Javascript/CSS are not loaded by the
browser.  A nicer approach is to use the Android Lint plugin to craft
the report in Jenkins.  Will be done in a follow up change.

Remove apps-android-wikipedia-lint
Remove Zuul specific patters
https://gerrit.wikimedia.org/r/#/c/345121/ does the switch in the source
repository.

Bug: T161305
Change-Id: Idfc181836ce38aecaca02407c50fe51f8e73d413