(WIP) ensure all jobs in Zuul are in JJB
Add jenkins-job-builder as a test dependency.
Generate all jobs in a temp directory.
Invoke zuul-server layout validation against the list of generated jobs.

That is similar to integration-zuul-layoutvalidation-gate which uses the
actual Jenkins to get the list of jobs.

Yield output such as:

FAILURE: Job analytics-libanon not defined
FAILURE: Job analytics-udp-filters not defined
FAILURE: Job analytics-webstatscollector not defined
FAILURE: Job analytics-wikistats not defined

Bug: T103847
Change-Id: I12917a96b554bca7a79c0fa1a80400f403a532f4