PoC: Port existing Trusty jobs to Docker
It would be nice to have a homogenous bunch of CI machines that only
need zuul, git, and docker-engine to run tests -- this patch is a
proof-of-concept towards that end.

This is a new job that does the same work as
mediawiki-phpunit-php55-trusty but it does so on a Debian Jessie Jenkins
agent node. This job is currently working, but the approach needs some
vetting and more eyes generally.

For this I re-used the integration-slave-docker-1000 box that marxarelli
built, I have some shell scripts in /srv/deployment/integration/docker
on that machine. Scripts are here:
https://github.com/thcipriani/integration-docker

Change-Id: Ied83cd30ad91f723d8174d754511139c88ba1161