Update patch set 2
Patch Set 2:

(3 comments)

That looks like a good first step for a simple jobs.  Eventually we would have to a reference Docker image that is build using the existing puppet manifest but that will be for later.

Can we provision a couple new instances dedicated for that on integration project ?  We would then probably want to check with ops about it.  Then I guess we can start migrating some of the simplest jobs.

Patch-set: 2