Update patch set 6
Patch Set 6:

Thanks for the review thcipriani! We're hoping to avoid having to maintain a separate repo at all, and just get the latest buildkit (i.e., the one Civi is using in their own CI) without having to regularly pull down updates.

It might be nice not to depend on GitHub, but unfortunately the buildkit already downloads a bunch of other tools from there. So unless we want to reinvent it, we're stuck with that GH dependency even if we use a local copy of buildkit. It used to be worse, though - an old version downloaded stuff from SourceForge!

Patch-set: 6
Reviewer: Ejegg <1070@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0