Clean out jsonlint/jshint
jsonlint and jshint are deprecated. They run a hardcoded version
installed directly on the CI slaves.  Moreover, we are in the process of
switching to ESLint which is done directly in the repos
package.json/Gruntfile.json.

For MediaWiki extensions:

Remove [jshint, jsonlint] from the check pipeline.

When mwgate-npm has been enabled, remove mwgate-jsonlint and
mwgate-jshint when grunt-jsonlint and grunt-contrib-jshint are present
in the package.json.

Also remove a few archived repositories while at it.

Bug: T119973
Bug: T154924
Change-Id: I66a4353c7c0f30ec548e49015d76542a5d509839