[SkelJS] drop jshint/jsonlint
The extension is 4+ years old and is supposed to propose a base skeleton
for JavaScript development.  It had little contributions and I want to
get rid of the jshint job from CI.

If there is interest in SkelJS again, it would be trivial to add a npm
job to it.

Change-Id: I1093dc7dfe67dcbdf487c3d474e403a9d1081e4a