Create SWAT gate and submit pipeline
This creates a new pipeline for wmf branches. Using branch regular
expressions with a negative lookahead allows you to duplicate the
gate and submit pipeline for all job templates and if the change is
being made to a wmf branch, then have that change be handled in a
different pipeline. This ensures that during SWAT changes to wmf
branches will fall into the gate-and-submit-swat pipeline while still
allowing changes to run as normal on the gate-and-submit pipeline.

Bug: T160668

Change-Id: I7f1ececaef4acf5729c8f3d9ba93c0617f27ca0c