Allow composer-test to run in different directory
To merge the composer-test functionality in mediawiki-testextension, we
need composer-test to run where the triggered repository has been cloned
to.  For an extension that would be src/extensions/$EXT_NAME.

Add a 'dir' parameter to the jjb macro 'composer-test'. That aligns it
with other macros 'composer-validate' and 'composer-update'.

Adjust the few callers to pass dir: '.'

Update:
    composer-hhvm-jessie
    composer-package-hhvm-jessie
    composer-package-php55-trusty
    composer-package-php70-jessie
    composer-php55-trusty
    composer-php70-jessie
    mwgate-composer-hhvm-jessie
    mwgate-composer-php55-trusty
    mwgate-composer-php70-jessie
    operations-mw-config-composer-hhvm-jessie
    translatewiki-composer-hhvm-jessie
    wikimedia-fundraising-crm-composer-php55-trusty

Bug: T161895
Change-Id: I32ff4e61fa9626599652ababafbefcf1de9693c1