Update patch set 1
Patch Set 1:

mwgate-composer-validate is applied by extension-unittests-generic . That is redundant when  mwgate-composer is applied for sure.

Then mwgate-composer-validate is running on permanent slaves, so at least it is not consuming an instance from Nodepool. Additionally it only triggers when a composer.json file is changed.


I have a patch to integrate composer validate/composer test in the testextension job https://gerrit.wikimedia.org/r/#/c/352160/   And I guess I will drop mwgate-composer-validate entirely as a result.

Patch-set: 1