Make mwext-MobileFrontend-npm-run-lint-modules job stable/voting
This job is not voting rather than non-voting to start with.
The reasoning is that if we make it non-voting we will not prioritise
fixes to the linter (a similar problem can be seen where browser
tests are not run on commits - they tend to stay broken for periods
of months). I've discussed this with Joaquin and Antoine and it's
trivial to revisit this decision and make this job non-voting if
we hit any problems.

Bug: T163882
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: Icd1905a81ab64bccf5cdc7f7f129685cfb9d9af7