Run composer test from mediawiki-extensions jobs
Just like we have merged 'composer test' in mwext-testextension jobs, do
the same for the mediawiki-extensions jobs (see 109ab65b93 ).

Zuul layout:
* Add phplint/composer validate in the check: pipeline.
* Remove mw-checks-test and mwgate-composer since that is now handled by
  running 'composer test'

Inject EXT_NAME when jobs are triggered from an extension. If set, the
job cd to the extension, run composer into it and clean up the
directory.  We do not want that when the jobs are triggered from
mediawiki/core or mediawiki/vendor,  use the Conditional build step
plugin to skip the run.

Bug: T161895
Change-Id: Ic960c80cf3901241408047df0a54eb5076e7a243