Docker for operations-puppet-tests
For future iterations, we can build out a series of base images that can
handle broad categories of job. This is very similar to how we spin up
build base images using Debian Image Builder for nodepool and should
require very little maintenance.

One part that doesn't scale with the approach in this patch is the
/run.sh file is very specific to this particular job since it runs both
tox and rake in parallel for speed. Ideally, we could figure out a way
to use the information we have in jjb to automatically generate a file
similar to this one, but this is a first pass.

The part that I do like about this approach is that it moves cloning the
git repository inside the container, so after the container has run
there is nothing left to clean up. This does rely on a bare reference
repository on the host at /srv/git/operations/puppet.git that is mounted
as a read-only volume in the container. We will need to update that repo
periodically.

The total runtime for a change in this instance is ~30 seconds since we
run both tox and rake tasks in parallel.

Bug: T166888
Change-Id: I0a06afd0dae162eb61508b9dd7ddf17f11690a68