Update patch set 2
Patch Set 2: Code-Review+1

As discussed as a first iteration that is good enough ™.  That opens a lot of questions for the future though:

- have a base docker image that already have apt.wm.o and its gpg key
- replace run.sh  with either jjb -> ansible or Makefile
- should the Dockerfile be directly in operations/puppet
- in which Docker registry to push the build container

etc.  But we can figure that out later on.

One strong requirement is to puppetize everything we need.  For ease of maintenance and ease of spawning more instances supporting Docker.

Patch-set: 2
Label: Code-Review=+1