Castor save operations/puppet postmerge
operations/puppet does not have any castor save running since there is
no gate-and-submit/postmerge pipeline for it.

Introduce postmerge job operations-puppet-cache-update-jessie which runs
with an empty cache, run the package manager installation steps and save
the result as if it got triggered by the job in test pipeline.

Only trigger it for 'production' branch and only if either Gemfile or
tox.ini are altered.

Have it send email to qa-alerts list on first failure (and abstract the
related code in a new JJB macro).

Bug: T168063
Change-Id: Ifc5985b0d9be0fd3cbe15d50c7e9655cd7a7c438