Move operations/puppet log volume
Because we were mounting the log volume at `.cache/log` and the `.cache`
directory didn't exist, that directory was created as the root user. As
a result, the process run inside the container by the `jenkins` user did
not have permission to create its needed cache directories.

This changes the mounted log volume to "$HOME/log" ($HOME is created as
part of the Dockerfile and has correct permissions). This also modifies
the Jenkins job slightly to use the new version of the container.

Change-Id: I7b522ac5e90f2e0af388c815d931a23251f4b37c