Update patch set 2
Patch Set 2:

We could +2 this CI config change and will effectively be merged as soon as the MinervaNeue tests are merged. However, the jobs are not published on the Jenkins instance automatically. That needs a manual step.

So here the Depends-On is to prevent us from deploying the jobs until the tests are available? :-)   Note:  they will not be available for production wikis until the tests make it to the wmf branch.

Patch-set: 2
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0