Update patch set 1
Patch Set 1:

We talked about that a while ago, this change is merely for demo purpose and is a NOOP.

The good thing with include-raw is you can do:

 - shell:
    !include-raw:
      - first.bash
      - second.bash

And both snippets would be in the same shell statement.

What that means is we can drop castor-define-namespace which is setup environment variables in both castor-load and castor-save and save them in an intermediate castor-set-namespace.env.  Since the snippets would run in the same shell, the intermediate file would be useless.

Patch-set: 1
Label: Verified=0