Update patch set 4
Patch Set 4:

Patch mostly seems fine aside from the extra job that is created. It is noteworthy that we'll have to add phpflavor-php56 labels to some nodes in order for the (mwgate-)/phplint56 jobs to work correctly.

I think the only reason that we don't run php55lint jobs on nodepool instances is that php5.5 is unavailable on jessie IIRC. The addition of those labels either means labeling permanent agents as phpflavor-php56 or figuring out if there is a way to add a label to nodepool instances when they register themselves with jenkins.

Patch-set: 4
Reviewer: Thcipriani <2321@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0