Update patch set 6
Patch Set 6: Code-Review+1

So that is surely need a bit of further polishing up:

* add a puppet class
* use disposable jenkins slaves (nodepool)
* add support to cache Rpackages

But that can be done as follow up patches I guess.  If this look good enough Mikhail, I will merge it tomorrow and then you will have the job run for patches to oritz (although it will always vote Verified+2).

Patch-set: 6