android: drop -no-audio
When not using the classic engine the emulator errors out with:

  qemu-system-i386: -audio: invalid option

The reason is QEMU2 replaced -audio with -soundhw. A workaround is to
configure QEMU2 via the environment variable: QEMU_AUDIO_DRV=none

Stop passing -no-audio to the emulator.
Add a job parameter QEMU_AUDIO_DRV=none

Note, we already:

  [android] Setting hardware properties:
    hw.audioInput: False

Reference:
https://stackoverflow.com/questions/40272884/running-android-emulator-with-noaudio-option-returns-qemu-system-i386-exe-au

Bug: T150623
Change-Id: I822057a37693735a1ccfeccc0fd23bb182c7b101