Update patch set 5
Patch Set 5: Code-Review-1

Those jobs are not cloning "mediawiki/extensions/DonationInterface/vendor"  but instead do a "composer install".

The behavior is defined in jjb/wm-fundraising.yaml in the macro "prepare-mediawiki-branch".


So we would need to pick some logic to select how the dependencies are installed.  Either from composer or from the vendor repo.


We did something similar for the mediawiki/services/parsoid . There is a job for each case and the logic to switch between is in jjb/macro.yaml "setup-npm-oid".

Or to say otherwise, the Parsoid jobs come in "source" and "deploy" flavors:

source: triggered on source repo, uses npm install

deploy: triggered on vendor repo, uses the vendor repo for dependencies


So it is a little bit more complicated.

Patch-set: 5
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1