Provide Android SDK location as an argument to non-periodic test scripts
The scripts for the non periodic Android CI jobs (apps-android-wikipedia-
test and apps-android-wikipedia-publish) have been updated in order to
handle an argument containing the location of an Android SDK installation.
[1] This will allow them to make use of a persisted Android SDK, rather than
re-downloading on every run as currently done, which is wasteful and slow.

This change knocks a minute off the time to completion for apps-android-
wikipedia-test, our primary CI testing job.

[1] https://gerrit.wikimedia.org/r/#/c/368118/

Bug: T171811
Change-Id: I993a3a60dd8ce9afd26dcbc2a0c253393cd7ec35