Update patch set 2
Patch Set 2:

I know Wikibugs sees it that way, but afaik this is not the case for Zuul somehow. I'm still having to remove my vote and re-+2 for it to happen.

Or maybe that's only the case for rebases? E.g. PS1 +2, PS2: Rebase, then it will require a re-submit manually (either by removing vote and re-adding, or maybe it will also work by just leaving any comment?)

Patch-set: 2