Update patch set 2
Patch Set 2: Code-Review+2

PY_COLORS=1 is to enable color output in tox when its running without a tty. tox 2.0.0 strip all env variables, so we gotta instruct it to accept the env.

I have also added it to a few places where it was missing.

Tyler: potentially the Docker image for puppet.git would have to be refreshed. Though it is not a priority.

Patch-set: 2
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2