Add Melos to trusted user list
While not an owner of a lot of patches, Melos is a long-term Wikimedian
and a steward, working recently on CentralAuth. Zuul running complete
tests on his patches will save us a lot of time doing "recheck"; and he
can be certainly trusted.

He's also a new mantainer of the labs/tools/stewardbots project with +2
permissions there. It only makes sense that his patches be fully tested
before he can merge them.

Change-Id: I8848244acb11f03c1ea01eb51634b859e6b0acd2