Update patch set 2
Patch Set 2:

PS2 actually switch cumin to use backports.


Based on a suggestion by Tyler on T17399 (hence the Signed-off-by). I went with aliases jobs that are exactly the same but named with '-backports' and whenever it is found in the job name, BACKPORTS=yes is injected.

Tyler suggested to use a Zuul tag, but that does not make things easier since we would still have to look at job.tags (instead of job.name) and would additionally need a rule in zuul such as:

 job:
  - name: ^debian-glue-backports
    tags: backports

Which is an extra rule being added.

Patch-set: 2