dib: allow tmpfs
If memory allow, dib would create two 4GB tmpfs to hold the build and
the image. That dramatically speed up the image creation.

I am not sure why I had it disabled. Maybe because it was originally
intended to run on labnodepool1001.eqiad.wmnet and would have required
more root access on the machine?

Change-Id: I90f13b1e90c852ae23b566782f12a2a9ff22fc4e