Marvin: install NPM >= 5
For stable builds, Marvin uses package-lock files which only NPM v5 and
greater support. Install a supporting NPM version before installing the
package dependencies.

This patch tries to keep things DRY but it fails in that respect a bit
and also has a side-effect change to every {name}-npm-node-6-jessie
job.

Prior to this patch:

  <hudson.tasks.Shell>
    <command>
      ...
      npm install
      npm test
    </command>
  </hudson.tasks.Shell>

After this patch:

  <hudson.tasks.Shell>
    <command>
      ...
    </command>
  </hudson.tasks.Shell>

  <hudson.tasks.Shell>
    <command>
      npm install
    </command>
  </hudson.tasks.Shell>

  <hudson.tasks.Shell>
    <command>
      npm test
    </command>
  </hudson.tasks.Shell>

The change is intended to be harmless but does occur quite ubiquitously.
This change has not been deployed.

Change-Id: I9f37d018f978005f1ed3e90e022bf0a81823b08c