dib: migrate out of <service>::packages
services::packages() is being phased out as services are moving to the
role/profile/module pattern.  The list of packages is now in the profile
which installs bunch of other materials.

Inline the packages definition.  Face a slight risk of being out of
sync, but they dont change that often.

Change-Id: I08d4bff0b77209cec5d0307a5d1d8b3b2279efb0