fabric: Add command to pull docker image on all hosts
Currently when we deploy a job with a new image or version
of an image the first job on each hosts can fail as the
image pull takes to long and when completing the job
itself has timed out.

This command can be used to pull a given tag of a given
image on all docker slaves before deploying the config
change to use the image.

It might be an idea to keep a list of all docker images
that are used & their version in the config repo that can
easily be fed into this command.
This also covers the case of starting up new docker slaves
and wanting it to have all relevant images (but I guess this
could also be in puppet?)

Change-Id: Ie75b1682111e637604f5018e653926003f196bc2