fabric: Add command to pull docker image in all hosts
Currently when we deploy a job with a new image or version
of an image the first job on each hosts can fail as the
image pull takes to long and when completing the job
itself has timed out.

This command can be used to pull a given tag of a given
image on all docker slaves before deploying the config
change to use the image.

Change-Id: Ie75b1682111e637604f5018e653926003f196bc2