Update patch set 1
Patch Set 1:

A new tidy version is up.

One thing to raise is that this removes the 3 apt packages needed to use sury in the single docker RUN command to save image space.
However this means that if a later image (such as the mediawiki specific php image) wants to install more things from sury (such as php exts) then it needs to re add the source, but I think that is okay...

Patch-set: 1